Move useFormControl into separate module
Making it more obvious that the full context is private
