+++ packages/material-ui/src/FormControl/useFormControl.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import FormControlContext from './FormControlContext';
+
+export default function useFormControl() {
+  return React.useContext(FormControlContext);
+}
